feat: add live drone map with react-leaflet and extract dashboard components
